requirements.txt
absl-py==1.4.0
astunparse==1.6.3
async-timeout==4.0.2
beautifulsoup4==4.12.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
deepface==0.0.79
filelock==3.10.2
fire==0.5.0
Flask==2.2.3
flatbuffers==23.3.3
gast==0.4.0
gdown==4.6.4
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
jax==0.4.6
Jinja2==3.1.2
keras==2.12.0
libclang==15.0.6.1
Markdown==3.4.2
MarkupSafe==2.1.2
mtcnn==0.1.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pip==23.0.1
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.3
requests==2.28.2
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
scipy==1.10.1
setuptools==65.6.3
six==1.16.0
soupsieve==2.4
spotipy==2.22.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wheel==0.38.4
wincertstore==0.2
wrapt==1.14.1
zipp==3.15.0
